requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
google-generativeai
google-cloud-firestore
spacy
pymupdf
matplotlib
pytest
httpx
python-multipart
python-dotenv
